- include qt 5 fixes - fixed bug in updateSubclassMediaType in mainw.cpp - updated readme - update isbn schema for correct parsing - updates spinbox range - removed warnings

+++ Programmierung/II/Praktikum/Beleg/manager/transactionman.cpp
@@ -160,7 +160,7 @@ int TransactionMan::removeTransactionId(unsigned long id) {
 
 // remove a transaction by index
 int TransactionMan::removeTransaction(unsigned long index) {
-    if (index < transactions.size()) {
+    if (index < (unsigned long) transactions.size()) {
         // get the transaction at the given index
         std::shared_ptr<Transaction> transaction = transactions[index];
         // remove the transaction from the vector
